+++ test/support/mock_index_worker.ex
@@ -11,6 +11,6 @@ defmodule TdDd.Search.MockIndexWorker do
   def reindex_count, do: Agent.get(MockIndexWorker, & &1)
 
   def reindex(_) do
-    Agent.update(MockIndexWorker, &[&1 + 1])
+    Agent.update(MockIndexWorker, &(&1 + 1))
   end
 end